ajusta as configurações do pwm

+++ pwm/pwm.h
@@ -3,7 +3,7 @@
 
 int pwm_enable();
 int pwm_disable();
-int pwm_period(int value);
+int pwm_period(float value);
 int pwm_duty(int value);
 
 #endif
